Split generate_data to gen_utils and generate_fer
